Eliminate all enemies ends level and melee attack is directiona
When all enemies are killed, the current level ends. All terrain and powerups are removed from the screen. The text at the top of the screen is changed to display level complete, and also how many tokens were acquired.

Melee attack is now directional. It attacks in the direction the mouse is facing. Animation is still a placeholder for now.