Shooter now has one life, game ends when dead
When shooters health gets to 0, it goes into STATE_DYING. From here, there is a time where the shooter does a death animation(still needs to be implemented), and then the shooter does to STATE_DONE.  When the player is dead, level counter is set back to 1, all enemies, terrain, and any attacks are cleared from the screen. The text at the top of the screen is set accordingly to let user know that he has died.